% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sum_contrasts.R
\name{nested_sum_contrasts}
\alias{nested_sum_contrasts}
\title{Sum contrasts for a nested factor}
\usage{
nested_sum_contrasts(outer.factor, inner.factor)
}
\arguments{
\item{outer.factor}{A factor.}

\item{inner.factor}{A factor whose levels are nested in the levels of the
outer.factor.}
}
\value{
An object of class \code{matrix} with \code{nlevels(inner.factor)}
rows and \code{nlevels(inner.factor) - nlevels(outer.factor)} columns.
}
\description{
The function creates sum contrasts for a factor whose levels are nested
within the levels of another factor.
}
\details{
The outer and inner factors need to have the same length. It can be data
that is used in a regression model, so it is allowed to have multiple
instances of the inner factor in the inner.factor, but it can also be
shorter factors that just describe the nesting.

The contrast matrix can be used directly on the inner factor using the
funtion contrasts, but it is usually necessary to specify the argument
how.many=nlevels(inner.factor) - nlevels(outer.factor).

The function \code{\link[psych]{superMatrix}} is used to make a block diagonal matrix.
}
\note{
The nesting of the input vectors is not checked.
}
\examples{


# Sum contrasts for the cancer type and the sample ID, which is nested in the cancer type.

data("cancermutations")

contrasts(cancermutations$cancer_type) = contr.sum(nlevels(cancermutations$cancer_type))
nesting = nested_sum_contrasts(outer.factor=cancermutations$cancer_type, inner.factor=cancermutations$sample_id)
how.many = nlevels(cancermutations$sample_id) - nlevels(cancermutations$cancer_type)
contrasts(cancermutations$sample_id, how.many=how.many) = nesting

# Visualize the contrast matrix: 
image(t(nesting[273:1,]))
}
\references{
Bertl, J.; Guo, Q.; Rasmussen, M. J.; Besenbacher, S; Nielsen, M. M.; Hornshøj, H.; Pedersen, J. S. & Hobolth, A. A Site Specific Model And Analysis Of The Neutral Somatic Mutation Rate In Whole-Genome Cancer Data. bioRxiv, 2017. doi: https://doi.org/10.1101/122879 \url{http://www.biorxiv.org/content/early/2017/06/21/122879}
}
\seealso{
\code{\link{contr.sum}}, \code{\link{contrasts}}, \code{\link{nested_treatment_contrasts}}
}
\author{
Johanna Bertl
}
\keyword{contrasts}
\keyword{regression}
