% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_multinom_dataprep.R
\name{wrapper_multinom_dataprep}
\alias{wrapper_multinom_dataprep}
\title{Wrapper for the preparation of mutation data for multinomial regression}
\usage{
wrapper_multinom_dataprep(cv, m, impute, datafolder, datafile, impfolder,
  data_source)
}
\arguments{
\item{cv}{integer. Which cross-validation piece should be handled?}

\item{m}{integer. Number of multiple imputation replicates. 0 means no imputation.}

\item{impute}{logical. Impute missing values or remove them?}

\item{datafolder}{character. The data is read from the file \code{paste0(datafolder, datafile, cv-1)}.}

\item{datafile}{character.}

\item{impfolder}{character. Folder where the imputed data should be saved.}

\item{data_source}{character. Either "fredriksson" or "pcawg".}
}
\value{
There is no return value. The reformatted and imputed data is saved along with the number of missing values and total mutation count.
}
\description{
This function is a wrapper for the function \code{\link{count_table_prep_multinom}} that prepares the mutation data for multinomial (and binomial) regression. This wrapper was used in our analysis in Bertl et al. (2007) (see References).
}
\details{
The scripts that were used to run this function and that show all settings used in Bertl et al. (2007) are available in this package in the folder \code{inst/Bertl_et_al_2017}. The pre-processed data can be downloaded from figshare.
}
\references{
Bertl, J.; Guo, Q.; Rasmussen, M. J.; Besenbacher, S; Nielsen, M. M.; Hornshøj, H.; Pedersen, J. S. & Hobolth, A. A Site Specific Model And Analysis Of The Neutral Somatic Mutation Rate In Whole-Genome Cancer Data. bioRxiv, 2017. doi: https://doi.org/10.1101/122879 \url{http://www.biorxiv.org/content/early/2017/06/21/122879}
}
\seealso{
\code{\link{count_table_prep_multinom}}, \code{\link{cancermutations}}
}
\author{
Johanna Bertl
}
