% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_prior_pseudo_counts.R
\name{wrapper_prior_pseudo_counts}
\alias{wrapper_prior_pseudo_counts}
\title{Wrapper to add prior pseudo counts to the mutation data}
\usage{
wrapper_prior_pseudo_counts(cv, mi, datafolder, dirichletfolder, rm_folder,
  rm_cancer = NULL)
}
\arguments{
\item{cv}{integer. Number of the cross-validation piece that should be handled.}

\item{mi}{integer. Number of the multiple imputation replicate that should be handled.}

\item{datafolder}{character. The dataset from the file \code{paste0(datafolder, "/", cv, "/imp", mi, ".txt")} is used.}

\item{dirichletfolder}{character. Where to save the dataset after the dirichlet counts were added.}

\item{rm_folder}{character. Where to save the dataset after the dirichlet counts were added and the cancer types (below) removed.}

\item{rm_cancer}{character vector. Cancer types that should be deleted (because they contain too few mutations) (or NULL).}
}
\value{
There is no return value. The data is saved in the specified folder(s).
}
\description{
This function is a wrapper for the function \code{\link{add_counts}} that adds prior dirichlet counts to the mutation data for multinomial (and binomial) regression. This wrapper was used in our analysis in Bertl et al. (2007) (see References).
}
\details{
The scripts that were used to run this function and that show all settings used in Bertl et al. (2007) are available in this package in the folder \code{inst/Bertl_et_al_2017}. The pre-processed data can be downloaded from figshare.
}
\references{
Bertl, J.; Guo, Q.; Rasmussen, M. J.; Besenbacher, S; Nielsen, M. M.; Hornshøj, H.; Pedersen, J. S. & Hobolth, A. A Site Specific Model And Analysis Of The Neutral Somatic Mutation Rate In Whole-Genome Cancer Data. bioRxiv, 2017. doi: https://doi.org/10.1101/122879 \url{http://www.biorxiv.org/content/early/2017/06/21/122879}
}
\seealso{
\code{\link{add_counts}}
}
\author{
Johanna Bertl
}
